add method isEnableToDelete in institutionService

+++ src/main/java/pl/coderslab/controllers/admin/InstitutionController.java
@@ -100,9 +100,13 @@ public class InstitutionController {
 
     @GetMapping("/delete/{id}")
     public String deleteInstitution(@PathVariable Long id){
-//      todo - sprawdzić czy nie ma powiązania z darami
-        institutionService.deleteInstitution(id);
+
+        if(institutionService.isEnableToDeleteInstitution(id)){
+            institutionService.deleteInstitution(id);
+            return "redirect:/admin/institutions/";
+        }
         return "redirect:/admin/institutions/";
+
     }
